requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
aiofiles
asyncpg
httpx
jwt
